requirements.txt
backports.zoneinfo==0.2.1
pkg_resources==0.0.0
plumbum==1.8.1
psycopg==3.1.8
psycopg-binary==3.1.8
typing_extensions==4.4.0
